feat(gpu): complete vectorized move semantics and add CI smoke test
GPU Move Application:
- Add departure marker placement on stack movement (RR-CANON-R090)
- Add own-marker path collapse with territory tracking
- Track eliminated_rings when landing on markers
- Fix capture to not count captured rings as eliminated (they're buried)

CI:
- Add gpu-smoke-test job to verify GPU module imports and basic ops

Performance:
- Pre-extract LPS data to CPU before loops to reduce .item() calls

+++ ai-service/app/training/value_calibration.py
@@ -30,11 +30,11 @@ class CalibrationBin:
 
     @property
     def mean_prediction(self) -> float:
-        return np.mean(self.predictions) if self.predictions else 0.0
+        return float(np.mean(self.predictions)) if self.predictions else 0.0
 
     @property
     def mean_outcome(self) -> float:
-        return np.mean(self.outcomes) if self.outcomes else 0.0
+        return float(np.mean(self.outcomes)) if self.outcomes else 0.0
 
     @property
     def calibration_error(self) -> float:
